Refactor + Clean up mqtt package (addresses #42)
