+++ server/handlers/deviceMessaging.go
@@ -3,11 +3,12 @@ package handlers
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/julienschmidt/httprouter"
-	"github.com/suyashkumar/conduit/server/mqtt"
 	"log"
 	"net/http"
 	"time"
+
+	"github.com/julienschmidt/httprouter"
+	"github.com/suyashkumar/conduit/server/mqtt"
 )
 
 type RpcResponse struct {
@@ -24,11 +25,11 @@ func Send(w http.ResponseWriter, r *http.Request, ps httprouter.Params, context 
 
 	prefixedName := PrefixedName(ps.ByName("deviceName"), hc.Prefix)
 
-	mqtt.SendMessage(prefixedName, ps.ByName("funcName"))
+	mqtt.Client().SendMessage(prefixedName, ps.ByName("funcName"))
 
-	c := make(chan string) 
+	c := make(chan string)
 
-	mqtt.Register(prefixedName+"/device", func(topic string, payload string) {
+	mqtt.Client().Register(prefixedName+"/device", func(topic string, payload string) {
 		defer func() {
 			if r := recover(); r != nil {
 				log.Println("Error in device response handler", r)
@@ -53,12 +54,12 @@ func Send(w http.ResponseWriter, r *http.Request, ps httprouter.Params, context 
 		fmt.Printf("Got device %s response", ps.ByName("deviceName"))
 	case <-timeout:
 		// Timed out!
-		SendErrorResponse(w, "ERROR, no response from device", 504) 
-		fmt.Println("Timeout waiting for response from device") 
+		SendErrorResponse(w, "ERROR, no response from device", 504)
+		fmt.Println("Timeout waiting for response from device")
 	}
 
-	// Cleanup: 
-	err := mqtt.DeRegister(prefixedName+"/device")
+	// Cleanup:
+	err := mqtt.Client().DeRegister(prefixedName + "/device")
 	if err != nil {
 		fmt.Println("Issues deregistering device message listener")
 	}
